+++ src/position.cpp
@@ -104,6 +104,40 @@ PositionBase PositionBase::make_move_copy(const MoveStr& move_str) const
     return make_move_copy(move);
 }
 
+void PositionNnue::accumulators_set(const Square sq, const Piece piece)
+{
+    EvaluationNnueBase::apply_piece<true>(accumulators, sq, piece);
+}
+
+void PositionNnue::accumulators_unset(const Square sq, const Piece piece)
+{
+    EvaluationNnueBase::apply_piece<false>(accumulators, sq, piece);
+}
+
+void PositionNnue::reset_accumulators()
+{
+    accumulators[Colors::White] = EvaluationNnueBase::input_biases;
+    accumulators[Colors::Black] = EvaluationNnueBase::input_biases;
+
+    for (Rank rank = 0; rank < 7; rank++)
+    {
+        for (File file = 0; file < 7; file++)
+        {
+            const auto sq = get_square(file, rank);
+            const auto piece = Squares[sq];
+
+            if (piece == Pieces::White)
+            {
+                accumulators_set(sq, piece);
+            }
+            else if (piece == Pieces::Black)
+            {
+                accumulators_set(sq, piece);
+            }
+        }
+    }
+}
+
 void Position::make_move_in_place(const Move& move)
 {
     assert(move != no_move);
@@ -129,6 +163,7 @@ void Position::make_move_in_place(const Move& move)
     Squares[move.To] = move.Turn;
     Bitboards[move.Turn] |= get_bitboard(move.To);
     Bitboards[Pieces::Empty] &= ~get_bitboard(move.To);
+    accumulators_set(move.To, move.Turn);
     Key ^= Zobrist.squares[move.Turn][move.To];
 
     // FROM
@@ -139,6 +174,7 @@ void Position::make_move_in_place(const Move& move)
         Squares[move.From] = Pieces::Empty;
         Bitboards[move.Turn] &= ~get_bitboard(move.From);
         Bitboards[Pieces::Empty] |= get_bitboard(move.From);
+        accumulators_unset(move.From, move.Turn);
         Key ^= Zobrist.squares[move.Turn][move.From];
     }
 
@@ -149,6 +185,8 @@ void Position::make_move_in_place(const Move& move)
     {
         const auto attacked_square = pop_lsb(attacked);
         Squares[attacked_square] = move.Turn;
+        accumulators_unset(attacked_square, !move.Turn);
+        accumulators_set(attacked_square, move.Turn);
         Key ^= Zobrist.squares[move.Turn][attacked_square];
         Key ^= Zobrist.squares[!move.Turn][attacked_square];
     }
@@ -184,6 +222,7 @@ void Position::unmake_move()
     Squares[move.To] = Pieces::Empty;
     Bitboards[move.Turn] &= ~get_bitboard(move.To);
     Bitboards[Pieces::Empty] |= get_bitboard(move.To);
+    accumulators_unset(move.To, move.Turn);
 
     // FROM
     if(undo_data.move.From != no_square)
@@ -193,6 +232,7 @@ void Position::unmake_move()
         Squares[move.From] = Turn;
         Bitboards[Turn] |= get_bitboard(move.From);
         Bitboards[Pieces::Empty] &= ~get_bitboard(move.From);
+        accumulators_set(move.From, move.Turn);
     }
 
     // CAPTURE
@@ -202,6 +242,8 @@ void Position::unmake_move()
     {
         const auto attacked_square = pop_lsb(undo_data.captured);
         Squares[attacked_square] = !move.Turn;
+        accumulators_unset(attacked_square, move.Turn);
+        accumulators_set(attacked_square, !move.Turn);
     }
     assert(verify_bitboards());
 }